[None][fix] clean DeepSeek V4 AutoDeploy ops
