+++ examples/auto_deploy/model_registry/configs/deepseek_v4_pr1_full.yaml
@@ -1,0 +1,62 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# DeepSeek V4 Flash PR1 full-layer config for generation quality validation.
+attn_backend: deepseek_v4_sparse
+compile_backend: torch-simple
+model_factory: DeepseekV4AutoModelForCausalLM
+skip_loading_weights: false
+max_batch_size: 1
+max_seq_len: 512
+max_num_tokens: 512
+enable_chunked_prefill: true
+cuda_graph_config: null
+model_kwargs:
+  skip_mtp: true
+  ad_rope_cache_len: 512
+  ad_compress_max_seq_len: 512
+transforms:
+  load_weights:
+    disable_preload: true
+  fuse_rmsnorm:
+    rmsnorm_backend: torch
+  detect_sharding:
+    enabled: false
+  sharding_transform_executor:
+    enabled: false
+  apply_sharding_hints:
+    enabled: true
+    dist_mapping:
+      tp: 8
+      moe_ep: 8
+      moe_tp: 1
+      moe_cluster: 1
+    enable_attention_dp: false
+    shard_layers: ["mla", "moe"]
+  insert_cached_attention:
+    backend: deepseek_v4_sparse
+  export_to_gm:
+    num_moe_experts_for_export: null
+  quantize_finegrained_fp8_linear_from_config:
+    enabled: true
+  fuse_finegrained_fp8_linear:
+    enabled: false
+  quantize_mxfp4_moe:
+    enabled: true
+  compile_model:
+    backend: torch-simple
+    piecewise_enabled: false
+  multi_stream_moe:
+    enabled: false
